update url source to use converters
