+++ packages/@atjson/source-url/package.json
@@ -13,9 +13,7 @@
   },
   "dependencies": {
     "@atjson/document": "file:../document",
-    "@atjson/offset-annotations": "file:../offset-annotations",
-    "@types/uuid": "^3.4.4",
-    "uuid": "^3.3.2"
+    "@atjson/offset-annotations": "file:../offset-annotations"
   },
   "devDependencies": {
     "@atson/renderer-commonmark": "file:../renderer-commonmark"
